% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{addKpredictions}
\alias{addKpredictions}
\title{addKpredictions}
\usage{
addKpredictions(whole, cv, new.clusters, indices)
}
\arguments{
\item{whole}{List. The output of Fluo_inspection() on the original data.}

\item{cv}{List. The output of Fluo_inspection() on the cross-validated data.}

\item{new.clusters}{Numeric. The predicted k-mean clusters.}

\item{indices}{List. The index numbers of the old and the new data.}
}
\value{
The new output of Fluo_inspection() with the original data where @GAPgroups has been replaced with
  the CV estimates and the new predictions and the @centroids with the CV-estimated centroids.
}
\description{
Adds the predicted k-mean clusters to the existing set (CV-estimated).
}
\keyword{internal}

