% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{aveDiff}
\alias{aveDiff}
\title{aveDiff}
\usage{
aveDiff(data, path.type, maxPseudo)
}
\arguments{
\item{data}{Numeric. A vector of pseudotimes whose first element is the originally estimated one (fromt he full data).}

\item{path.type}{Character. The input of path.type parameter in pathEstimator().}

\item{maxPseudo}{Numeric. The maximum possible pseudotime.}
}
\value{
The average difference between the original and the CV estimated pseudotimes for a sample.
}
\description{
It estimates the average difference between the original and the CV estimated pseudotimes. For circular path
  types, the difference is defined as min(diff,max(pseudotime)-diff). For example assuming 300 cells (thus the
  maximum pseudotime is 300) in a circular path, two pseudotimes 1 and 300 differ only by 1 and not by 299.
}
\keyword{internal}

