% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{BGcorrectFluo}
\alias{BGcorrectFluo}
\title{BGcorrectFluo}
\usage{
BGcorrectFluo(data, method, old.offset, bg)
}
\arguments{
\item{data}{Data matrix. A data matrix with the foreground and background raw signals
from each channel.}

\item{method}{Character string. The type of background correction to be performed. One
of "normexp" or "subtract".}

\item{old.offset}{Float. An offset for the background correction method.}

\item{bg}{Logical. If TRUE foreground - background is performed.}
}
\value{
The background corrected data
}
\description{
It performs the background correction of cell fluorescence signals.
}
\keyword{internal}

