% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{boxcoxMatrix}
\alias{boxcoxMatrix}
\title{boxcoxMatrix}
\usage{
boxcoxMatrix(data)
}
\arguments{
\item{data}{Numeric vector. The adjusted signals of one channel.}
}
\value{
The transformed data
}
\description{
A helper to run the Box Cox transformation on a data matrix of adjusted fluorescence signals.
}
\keyword{internal}

