% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{boxFluo}
\alias{boxFluo}
\title{boxFluo}
\usage{
boxFluo(data, transformation, reference, legends, batchnames, image.type,
  savePlot)
}
\arguments{
\item{data}{List. A list with the fluorescence data. Typically, the output of createFluo().}

\item{reference}{Integer. The run number to be used as baseline for the run correction.}

\item{legends}{Character vector. Puts the "uncorrected" or "corrected" legends on the signal density
plots.}

\item{batchnames}{original run name.}

\item{image.type}{Character string. A triplet of IDs to characterize the type of images under study.}

\item{savePlot}{Character. A switch to generate the density plots.}

\item{transform}{Character string. The type of transformation to be performed. One
of "bc" (Box-Cox), "log", "log10" or "asinh".}
}
\value{
The density plots of the fluorescence data
}
\description{
It generates the density plots of the uncorrected and corrected cell fluorescence signals.
}
\keyword{internal}

