% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{contrastFluo}
\alias{contrastFluo}
\title{contrastFluo}
\usage{
contrastFluo(data, channel, legends)
}
\arguments{
\item{data}{List. A list with the fluorescence signal information of both channels. Typically,
the output of adjustFluo().}

\item{channel}{Character string. An identifier for the channel to be analyzed.}
}
\value{
A list with the fluorescence signals, mixture components, flexmix model estimates and
  contrast results
}
\description{
It estimates the contrasts comparisons across runs and runs*component in each channel.
}
\keyword{internal}

