% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{cpPELT}
\alias{cpPELT}
\title{cpPELT}
\usage{
cpPELT(data, sig.level, thresh, seed)
}
\arguments{
\item{data}{List. A list of adjusted fluorescence signals in both channels. Typically,
the output of transformFluo().}

\item{sig.level}{Float. The significance level below which we do not reject a change point.}

\item{thresh}{Integer. The minimum number of values for a cluster re-estimated by the
change-point analysis.}

\item{seed}{Integer. An optional seed number for the Random Number Generator.}
}
\value{
A list of change-points and the associated change-point clusters
}
\description{
It performs the change-point analysis of the variance stabilized adjusted fluorescence signals by PELT.
}
\keyword{internal}

