% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{cutpointsEstimator}
\alias{cutpointsEstimator}
\title{cutpointsEstimator}
\usage{
cutpointsEstimator(data, thresh, cmethod, sig.level, Q, seed)
}
\arguments{
\item{data}{List. A list of adjusted fluorescence signals for both channels.}

\item{thresh}{Integer. The minimum number of values for a cluster re-estimated by the
change-point analysis.}

\item{cmethod}{Character string. The change point method to be used. It can be one of "ECP",
(non-parametric) "manualECP" (non-parametric with user-defined numner of change-points) or
"PELT" (Pruned Exact Linear Time; parametric).}

\item{sig.level}{Float. The significance level below which we do not reject a change point.}

\item{Q}{Integer. The number of change-points to be kept if CPmethod = "manualECP".}

\item{seed}{Integer. An optional seed number for the Random Number Generator.}
}
\value{
The sorted transformed signal differences (path) and the associated change-points
}
\description{
A helper that estimates the change-points by various methods.
}
\keyword{internal}

