% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{DDHFfit}
\alias{DDHFfit}
\title{DDHFfit}
\usage{
DDHFfit(data, den.method, savePlot)
}
\arguments{
\item{data}{List. A list of fluorescence signals with their change-points and clusters.}

\item{den.method}{Character string. A method to perform the denoising. One of "wavelets", "splines"
or "lregr" (for linear regression).}

\item{savePlot}{Character string. The directory to store the plots.}
}
\value{
The DDHFmv clusters and model estimates
}
\description{
An internal function to produce the DDHFmv clustering and other model estimates.
}
\keyword{internal}

