% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{DDHFinput}
\alias{DDHFinput}
\title{DDHFinput}
\usage{
DDHFinput(data, ms)
}
\arguments{
\item{data}{Data matrix. A data matrix of fluorescence signals.}

\item{ms}{Data matrix. A matrix of estimated cluster centroids.}
}
\value{
The sorted fluorescence signals
}
\description{
It sorts the fluorescence signals of both channels data for DDHF.
}
\keyword{internal}

