% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{ddhft.np.2}
\alias{ddhft.np.2}
\title{ddhft.np.2}
\usage{
ddhft.np.2(data)
}
\arguments{
\item{data}{Numeric vector. A vector of data exhibiting monotonically increasing
mean-variance relationship. The data will be transformed.}
}
\value{
The DDHF transformed data
}
\description{
The original DDHF function (Motakis et al, 2006).
}
\keyword{internal}

