% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{denoisefun}
\alias{denoisefun}
\title{denoisefun}
\usage{
denoisefun(data, method)
}
\arguments{
\item{data}{Numeric vector. An 1-dimensional vector of pseudotime sorted transformed signal differences.}

\item{method}{Character string. A method to perform the denoising. One of "wavelets", "splines" or "lregr"
(linear rergression).}
}
\value{
The denoised data
}
\description{
An internal function to run data denoising.
}
\keyword{internal}

