% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{denoiser}
\alias{denoiser}
\title{denoiser}
\usage{
denoiser(data, method)
}
\arguments{
\item{data}{Data matrix. A matrix contains the sample indices, the pseudotimes, the Ch2-Ch3 transformed data
and the cluster numbers.}

\item{method}{Character string. A method to perform the denoising. One of "wavelets", "splines" or "lregr"
(linear regression).}
}
\value{
The denoised data and the model residuals
}
\description{
An internal function to run transformed data denoising that estimates the model residuals.
}
\keyword{internal}

