% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{despecklefun}
\alias{despecklefun}
\title{despecklefun}
\usage{
despecklefun(img, pix, thresh)
}
\arguments{
\item{img}{Data matrix. The BF image data matrix.}

\item{pix}{Integer. A cutoff specifing the area to be examined for speckles.}

\item{thresh}{Integer. A cutoff to perform the despeckle function. If pixel signal > median object signal + thresh,
the object is a speckle and the median object signal is returned.}
}
\value{
A despeckled image
}
\description{
It despeckles the BF image data matrix (similar to the despeckle function of ImageJ)
}
\keyword{internal}

