% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{diagnoseResiduals}
\alias{diagnoseResiduals}
\title{diagnoseResiduals}
\usage{
diagnoseResiduals(data, savePlot = "OFF")
}
\arguments{
\item{data}{List. A list of fluorescence signals with their model estimates.}

\item{savePlot}{Character string. The directory to store the plots.}
}
\value{
The residual diagnostics and plots
}
\description{
An internal function to perform residual diagnostic tests.
}
\keyword{internal}

