% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{distfromcenter}
\alias{distfromcenter}
\title{distfromcenter}
\usage{
distfromcenter(data, center)
}
\arguments{
\item{data}{Data matrix. A 2-dimensional location.}

\item{center}{Data matrix. A second 2-dimensional location.}
}
\value{
The Eucledian distance between the two locations
}
\description{
It calculates the Eucledian distance between two 2-dimensional locations.
}
\keyword{internal}

