% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{estimate.new.pseudotimes}
\alias{estimate.new.pseudotimes}
\title{estimate.new.pseudotimes}
\usage{
estimate.new.pseudotimes(data)
}
\arguments{
\item{data}{Data matrix. The estimated pseudotimes for all samples with the original data and the CV data undet two methods,
i.e. "median/original" and median/null". For details see parameter pseudo.est.method at Fluo_CV_modeling().}
}
\value{
It summarizes the CV-estimated pseudotimes into a single value. There are three possible methods that may produce
  different results. For details see parameter pseudo.est.method at Fluo_CV_modeling().
}
\description{
It estimates a unique pseudotime vector to be used for analysis with Fluo_ordering(). It is based on the cross-validation
  estimates for each sample and a particular method defined by pseudo.est.method parameter at Fluo_CV_modeling().
}
\keyword{internal}

