% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{estimatePath}
\alias{estimatePath}
\title{estimatePath}
\usage{
estimatePath(data, type, start)
}
\arguments{
\item{data}{Data matrix. A matrix of centroids with their trigonometric function values.}

\item{type}{Character string. A user-defined value that characterizes the cell progression dynamics.
It can be either "clockwise" or "anticlockwise" depending on how the path is expected
to proceed.}

\item{start}{Integer. The cluster number that is assigned as the path starting point}
}
\value{
The sorted cluster indices (path)
}
\description{
The main function for automatic path estimation .
}
\keyword{internal}

