% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{extractBFArea}
\alias{extractBFArea}
\title{extractBFarea}
\usage{
extractBFArea(cc, img, area, BFarea)
}
\arguments{
\item{cc}{Data matrix. An estimated spot center.}

\item{img}{Data matrix. The matrix of image data from a channel.}

\item{area}{Data matrix. The bright (spot) coordinates around the spot center.}

\item{BFarea}{Integer. Defines a rectangular pseudo-spot size whose fluorescence will be estimated.}
}
\value{
The area identified by BF image modelling
}
\description{
It estimates the spot or capture site location by BF image modelling.
}
\keyword{internal}

