% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{failurecase}
\alias{failurecase}
\title{failurecase}
\usage{
failurecase(img, pattern.search, despeckle, ImgLimits, chip.type, separator,
  image.type)
}
\arguments{
\item{pattern.search}{Integer. A cutoff to find horizontal and vertical lines on the chip.}

\item{despeckle}{Logical. If TRUE, the BF image is descpeckled.}

\item{ImgLimits}{Integer. It instructs the algorithm to find spots in a certain central image area.
For example, for a 512 x 512 image with ImgLimits = 50, it will search for spots in
the central area [ImgLimits:(512-ImgLimits),ImgLimits:(512-ImgLimits)] of the image matrix.}

\item{separator}{Character string. Removes the Bright Field ("BF") and channel indicators from the
image file names.}

\item{image.type}{Character string. A triplet of IDs to characterize the type of images under study.}

\item{origImg}{Data matrix. The original BF image to be read and processed.}

\item{type}{Character string. It specifies the type of Fluidigm chip to be analyzed.}
}
\value{
Location statistics and characteristic lines of BF image modelling
}
\description{
Another helper to re-estimate the spot location or find the capture site location
  (BF image modelling).
}
\keyword{internal}

