% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{fclust}
\alias{fclust}
\title{fclust}
\usage{
fclust(data, k, nstart = 1)
}
\arguments{
\item{data}{Data matrix. A matrix of run effect and background corrected fluorescence signals
in both channels.}

\item{k}{Integer. The maximum number of clusters to be generated.}

\item{nstart}{Integer. A flowClust parameter specifing the number of random sets to
be chosen for the clustering estimation.}
}
\value{
The flowClust results
}
\description{
The main flowClust function used in this application.
}
\keyword{internal}

