% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{filled.contour3}
\alias{filled.contour3}
\title{filled.contour3}
\usage{
filled.contour3(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1,
  length.out = ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y,
  finite = TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim,
  nlevels), nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), joinedPlots, plot.title, plot.axes,
  key.title, key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1,
  axes = TRUE, frame.plot = axes, mar, ...)
}
\arguments{
\item{x, y, z}{Numeric vectors. Some plot coordinates.}

\item{...}{Other parameters of the function.}
}
\value{
A plotted spot
}
\description{
It generates a contour plot of a channel specific spot. It is a modification by Ian Taylor of the filled.contour()
  to remove the key and facilitate overplotting with contour(). It has been further modified by Carey McGilliard and
  Bridget Ferris to allow multiple plots on one page http://wiki.cbr.washington.edu/qerm/sites/qerm/images/1/16/Filled.contour3.R
  We have added some extra parameters to adapt the function to our application.
}
\keyword{internal}

