% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{findPattern}
\alias{findPattern}
\title{findPattern}
\usage{
findPattern(imgline, bpattern)
}
\arguments{
\item{imgline}{Integer. A row or a column number of the BF image data matrix.}

\item{bpattern}{Integer. A user-defined cutoff that specifies whether a given row
or column contains a characteristic straight line.}
}
\value{
An estimate for the existence of a characteristic line
}
\description{
A helper to find the characteristic straight lines of the BF image .
}
\keyword{internal}

