% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{fixPath}
\alias{fixPath}
\title{FixPath}
\usage{
fixPath(data, groups)
}
\arguments{
\item{data}{List. A list of fluorescence signal information for both channels.}

\item{groups}{Numeric vector. A vector of cluster indices.}
}
\value{
A list with the adjusted fluorescence signals and the clusters
}
\description{
It tests whether the path has been appropariately defined and produces an error if not.
}
\keyword{internal}

