% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{flowclust_step1}
\alias{flowclust_step1}
\title{flowclust_step1}
\usage{
flowclust_step1(data, k, nstart)
}
\arguments{
\item{data}{Data matrix. A matrix of run effect and background corrected fluorescence signals
in both channels.}

\item{k}{Integer. The maximum number of clusters to be generated.}

\item{nstart}{Integer. A flowClust parameter specifing the number of random sets to
be chosen for the clustering estimation.}
}
\value{
Preliminary flowClust results
}
\description{
A helper function for flowClust analysis.
}
\keyword{internal}

