% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{flowclust_step2}
\alias{flowclust_step2}
\title{flowclust_step2}
\usage{
flowclust_step2(data)
}
\arguments{
\item{data}{Data matrix. A matrix of flowclust results from flowclust_step1().}
}
\value{
Preliminary flowClust results
}
\description{
Another helper function for flowClust analysis.
}
\keyword{internal}

