% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{FluoInspection}
\alias{FluoInspection}
\title{FluoInspection}
\usage{
FluoInspection(data, dateIndex, savePlot)
}
\arguments{
\item{data}{List. A list of fluorescence signal information for both channels.}

\item{dateIndex}{Character string. A date index to be used in saving the output files.}

\item{savePlot}{Character string. The directory to store the plots or an option to print
them on the screen.}
}
\value{
A list with the adjusted fluorescence signals and the centroids
}
\description{
It generates the initial clusters, their centroids and plots the results.
}
\keyword{internal}

