% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{forceBF}
\alias{forceBF}
\title{forceBF}
\usage{
forceBF(data, cutoff, median.correction, medians, Wells, image.type)
}
\arguments{
\item{data}{List. The location statistics of both channels.}

\item{cutoff}{Integer. A cutoff do detect outlier locations.}

\item{median.correction}{Logical. If TRUE it performs median adjustment for outlier locations.}

\item{medians}{Data matrix. The estimated medians of non-outlier samples by run and well ID.}

\item{Wells}{Data matrix. The directionality of the well IDs.}

\item{image.type}{Character string. A triplet of IDs to characterize the type of images under study.
They refer to the ImageType part of the original image or txt file names.}
}
\value{
A list of re-estimated locations
}
\description{
It re-estimates the location of the outlier samples
}
\keyword{internal}

