% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{function.from.vector}
\alias{function.from.vector}
\title{function.from.vector}
\usage{
function.from.vector(x, y, argument.vect)
}
\arguments{
\item{x, y, argument.vector}{Appropriate vectors for analysis.}
}
\value{
Preliminary DDHF results
}
\description{
A helper for DDHF.
}
\keyword{internal}

