% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{getCoordinates_stats}
\alias{getCoordinates_stats}
\title{getCoordinates_stats}
\usage{
getCoordinates_stats(centerR, centerG, minDiff, chaImgs, ll, ws, estCenter)
}
\arguments{
\item{centerR}{Data matrix. The location statistics in one channel.}

\item{centerG}{Data matrix. The location statistics in thew other channel.}

\item{minDiff}{Float. the mu_hat of the H0: image-to-noise ratio =
log(foreground_signal) - log(background_signal) = mu_hat. Rejection of H0
implies that the identified spot is brighter than background. Default is 0.5.}

\item{chaImgs}{List. A list of the channel images (data matrices) of a sample.}

\item{ll}{Data matrix. An internal parameter specifying the spot center.}

\item{ws}{List. An internal parameter specifying the spot center.}

\item{estCenter}{Data matrix. The estimated spot center by BF image modelling.}
}
\value{
A series of location estimates includying the channel-specific spot center and spot areas
}
\description{
It finds the spot coordinates using the spot centers or BF image modeling
}
\keyword{internal}

