% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{getSpot}
\alias{getSpot}
\title{getSpot}
\usage{
getSpot(img, rad)
}
\arguments{
\item{img}{Data matrix. The binary segmented channel image data.}

\item{rad}{Integer. A cut-off to separate the spots (densely clustered 1s) from small random signals
(loosely clustered 1s) on the binary image.}
}
\value{
A matrix of bright (spot) coordinates
}
\description{
Identifies one or multiple spot(s) in the image data matrix.
}
\keyword{internal}

