% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{giveWarning}
\alias{giveWarning}
\title{giveWarning}
\usage{
giveWarning(number)
}
\arguments{
\item{number}{Integer. An internally defined number that produces a message .}
}
\value{
A coded message
}
\description{
It generates a coded message of the estimation type that is being performed.
}
\keyword{internal}

