% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{grouplines}
\alias{grouplines}
\title{grouplines}
\usage{
grouplines(data)
}
\arguments{
\item{data}{Integer. A row or a column of the BF image data matrix.}
}
\value{
An estimate for the existence of a characteristic line
}
\description{
It reconstructs the BF image characteristic lines.
}
\keyword{internal}

