% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{grubbs}
\alias{grubbs}
\title{grubbs}
\usage{
grubbs(data)
}
\arguments{
\item{data}{Numeric vector. An 1-dimensional vector of spot distances to check for outliers.}
}
\value{
All potential outliers (indices)
}
\description{
It performs the grubbs test for outliers.
}
\keyword{internal}

