% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{GrubbsOutliers}
\alias{GrubbsOutliers}
\title{GrubbsOutliers}
\usage{
GrubbsOutliers(data, alpha)
}
\arguments{
\item{data}{Data matrix. A data matrix of fluorescence signals and model residuals.}

\item{alpha}{Float. A significance level for the grubbs test .}
}
\value{
The fluorescence signals and the potential outliers
}
\description{
It identifies potential outliers by the Grubbs test.
}
\keyword{internal}

