% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{highlight.cols}
\alias{highlight.cols}
\title{highlight.cols}
\usage{
highlight.cols(data, chip.type, fac)
}
\arguments{
\item{data}{Numeric. A column of the BF image data matrix.}

\item{chip.type}{Character string. It specifies the type of Fluidigm chip to be analyzed.}

\item{fac}{Float. An internally estimated factor that adjusts for different chip types.}
}
\value{
Estimated vertical BF characteristic lines.
}
\description{
A helper to identify the vertical BF image characteristic lines.
}
\keyword{internal}

