% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{invTransform}
\alias{invTransform}
\title{invTransform}
\usage{
invTransform(data, lambda, transformation)
}
\arguments{
\item{data}{Data matrix. The adjusted signals of both channels.}

\item{lambda}{Float. The lambda parameter of the Box-Cox.}

\item{transform}{Character string. The type of transformation to be performed. One
of "bc" (Box-Cox), "log", "log10" or "asinh".}
}
\value{
The back-transformed data matrix
}
\description{
It back-transforms the transformed adjusted cell fluorescence signals of a matrix.
}
\keyword{internal}

