% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{isotone}
\alias{isotone}
\title{isotone}
\usage{
isotone(x, wt = rep(1, length(x)), increasing = TRUE)
}
\arguments{
\item{x}{Numeric vector. A vector of appropriately sorted data.}

\item{...}{Other parameters.}
}
\value{
The isotone regression model estimates
}
\description{
The original function to perform isotone regression (Motakis et al 2006).
}
\keyword{internal}

