% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{joinAreas}
\alias{joinAreas}
\title{joinAreas}
\usage{
joinAreas(areaR, areaG, center, chaImg, areaBased, warning)
}
\arguments{
\item{areaR}{Data matrix. The bright (spot) coordinates in one channel.}

\item{areaG}{Data matrix. The bright (spot) coordinates in the other channel.}

\item{center}{Data matrix. The 2-dimensional location of the spot's center.}

\item{areaBased}{Integer. Defines a rectangular pseudo-spot size whose fluorescence will be estimated.}

\item{warning}{Character string. An indicator of the estimation type that has been internally performed, i.e.
fluorescence-based or BF image modelling.}

\item{chaImgs}{Data matrix. The channel binary segmented image data that are used for reference to obtain the
area of interest.}
}
\value{
The coordinates of the spot area and its length in pixels.
}
\description{
It estimates the spot area by joining the red and green bright spot location estimates.
}
\keyword{internal}

