% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{listSorter}
\alias{listSorter}
\title{listSorter}
\usage{
listSorter(data, sorter)
}
\arguments{
\item{data}{List. A list variable.}

\item{sorter}{Numeric vector. An appropriate sorter.}
}
\value{
A list of appropriately sorted data
}
\description{
A helper that sorts the data of a list variable.
}
\keyword{internal}

