% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{lmFluo}
\alias{lmFluo}
\title{lmFluo}
\usage{
lmFluo(data, batch, maxMix, reference, prior.pi, flex.reps, flexmethod, seed)
}
\arguments{
\item{data}{Numeric vector. An 1-dimensional vector of adjusted data from a single channel.}

\item{batch}{Integer. The run number.}

\item{maxMix}{Integer. The maximum number of components to fit into the mixture of
regressions model.}

\item{reference}{Numeric vector. Specifies the runs to be used as baseline (iteratively).}

\item{prior.pi}{Float. The prior probability to accept a component.}

\item{flex.reps}{Integer. The iterations of the Expectation-Maximization algorithm to estimate
the flexmix model.}

\item{flexmethod}{Character string. A method to estimate the optimal number of flexmix
components. One of "BIC", "AIC", "ICL".}

\item{seed}{Integer. An optional seed number for the Random Number Generator.}
}
\value{
The flexmix mixture components and other statistics
}
\description{
It estimates the optimal number of mixtures for the flexmix model on data from multiple runs.
}
\keyword{internal}

