% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{mean_signal}
\alias{mean_signal}
\title{mean_signal}
\usage{
mean_signal(data, noise.level)
}
\arguments{
\item{data}{Numeric vector. An 1-dimensional vector of spot signals.}

\item{noise.level}{Float. The noise level of the image.}
}
\value{
The average spot's noisy signal
}
\description{
A helper to simulate the spot signal.
}
\keyword{internal}

