% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{measureB}
\alias{measureB}
\title{measureB}
\usage{
measureB(img, area, iter, BFarea)
}
\arguments{
\item{img}{Data matrix. The matrix of image data from a channel.}

\item{area}{Data matrix. The bright (spot) coordinates around the spot center.}

\item{iter}{Integer. A number of iterations (pseudo-spots) to be summarized.}

\item{BFarea}{Integer. Defines a rectangular pseudo-spot size whose fluorescence will be estimated.}
}
\value{
The background image estimates
}
\description{
It estimates the background signal for an image.
}
\keyword{internal}

