% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{myt}
\alias{myt}
\title{myt}
\usage{
myt(data, minDiff)
}
\arguments{
\item{data}{Numeric vector. The 1-dimensinal signal of log(foreground) - log(background).}

\item{minDiff}{Float. The mu_hat of the H0: image-to-noise ratio =
log(foreground_signal) - log(background_signal) = mu_hat. Rejection of H0
implies that the identified spot is brighter than background.}
}
\value{
A test P-value
}
\description{
A helper to test whether the foreground signal is statistically higher than the background.
}
\keyword{internal}

