% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{orderFluo}
\alias{orderFluo}
\title{orderFluo}
\usage{
orderFluo(data, path.type, updater = FALSE)
}
\arguments{
\item{data}{List. A list of adjusted fluorescence signals.}

\item{updater}{Logical. An indicator of the estimation stage. If FALSE the initial groups are analyzed,
otherwise the change-point based groups are analyzed.}

\item{path.start}{Integer. A cluster number indicating the starting cluster that algorithm should use to
build the path. The cluster numbers refer to the plot generated by Fluo_inspection(). Default is 1.
If path.type = "circular" the number does not matter. If path.type = "A2Z" the user should inspect the
Fluo_inspection() plot to detect the beginning of the path. If path.type = "other", the function will
not estimate a path. The user has to manually insert the path progression (the cluster numbers) in
Fluo_modeling()  .}
}
\value{
A list with the adjusted fluorescence signals, the centroids, the clusters and the pseudotimes
}
\description{
It sort the adjusted (and transfomed) fluorescence signals according to the path progression.
}
\keyword{internal}

