% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{path.initiator}
\alias{path.initiator}
\title{path.initiator}
\usage{
path.initiator(data, where)
}
\arguments{
\item{data}{List. The output of Fluo_inspection().}

\item{where}{Character. One of "bottom/left", "bottom/right", "top/left", "top/right" that specify
the starting point of the progression path.}
}
\value{
A starting point for the progression path
}
\description{
It finds the cluster that initiates the progression path.
}
\keyword{internal}

