% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{pathUpdater}
\alias{pathUpdater}
\title{pathUpdater}
\usage{
pathUpdater(data, path)
}
\arguments{
\item{data}{Data matrix. A matrix of centroids with their path progression indices.}

\item{path}{Numeric vector. The path progression indices.}
}
\value{
The sorted cluster indices (path)
}
\description{
A helper that updates the path sorted clusters after re-estimation by change-point analysis.
}
\keyword{internal}

