% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{plot_clusgap}
\alias{plot_clusgap}
\title{plot_clusgap}
\usage{
plot_clusgap(clusgap, title = "Gap Statistic results")
}
\value{
plot gap stats
}
\description{
adapted from plot_clusgap from phyloseq.
}
\keyword{internal}

