% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{predict.kmeans}
\alias{predict.kmeans}
\title{predict.kmeans}
\usage{
\method{predict}{kmeans}(data, centroid)
}
\arguments{
\item{data}{Data matrix. A 2-dimensional matrix of corrected fluorescence signals that are clustered by k-means.}

\item{centroid}{Data matrix. A 2-dimensional matrix of the k-means centroids.}
}
\value{
The predicted k-mean clusters.
}
\description{
Takes a training sample and predicts the k-mean clusters of a new dataset (minimizing the Eucledian distance).
}
\keyword{internal}

