% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{processImg}
\alias{processImg}
\title{processImg}
\usage{
processImg(img, denoise)
}
\arguments{
\item{img}{Data matrix. The matrix of image data from a channel.}

\item{denoise}{Logical. If TRUE the channel image is denoised with 2-dimensional la8, universal
and hard thresholding.}
}
\value{
A denoised image
}
\description{
Denoises the image data to be used for spot location estimation.
}
\keyword{internal}

