% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{project}
\alias{project}
\title{project}
\usage{
project(data, centers)
}
\arguments{
\item{data}{Data matrix. The fluorescence signals of a particular cluster in both channels.}

\item{centers}{Numeric vector. A vector of the 2-dimensional cluster centroids.}
}
\value{
The pseudotimes and cell progression
}
\description{
A helper to estimate the pseudotimes by orthogonal projection.
}
\keyword{internal}

