% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{readChaImg}
\alias{readChaImg}
\title{readChaImg}
\usage{
readChaImg(imgNames, denoise)
}
\arguments{
\item{denoise}{Logical. If TRUE the channel image is denoised with 2-dimesnional la8, universal
and hard thresholding.}

\item{imgName}{Character string. The file name of the images.}
}
\value{
A list of channel image estimates
}
\description{
It reads and processes the channel image data
}
\keyword{internal}

