% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{readOriImg}
\alias{readOriImg}
\title{readOriImg}
\usage{
readOriImg(imgName, despeckle, pix, thresh, separator, image.type)
}
\arguments{
\item{imgName}{Character string. The file name of the image.}

\item{despeckle}{Logical. If TRUE the BF image is despeckled.}

\item{pix}{Integer. A cutoff specifing the size of the area to search for speckles.}

\item{thresh}{Integer. A cutoff to perform the despeckle function. If pixel signal > median object signal + thresh,
the object is a speckle and the median object signal is returned.}

\item{separator}{Character string. Removes the Bright Field ("BF") and channel indicators from the
image file names.}

\item{image.type}{Character string. A triplet of IDs to characterize the type of images under study. They refer to
the ImageType part of the original image or txt file names.}
}
\value{
A list of BF image estimates
}
\description{
It reads and processes the original BF image
}
\keyword{internal}

