% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{refineMixes}
\alias{refineMixes}
\title{refineMixes}
\usage{
refineMixes(data, batch, model, seed)
}
\arguments{
\item{data}{Numeric vector. The background corrected fluorescence signals of
a single channel.}

\item{batch}{Integer. The run number.}

\item{model}{Model. The output of getModel() from flexmix.}

\item{seed}{Integer. An optional seed number for the Random Number Generator.}
}
\value{
The flexmix mixture components
}
\description{
An helper internal function to generate the results of flexmix.
}
\keyword{internal}

