\name{Results}
\alias{Results}
\docType{data}
\title{
Results
}
\description{
Example output from LocationMatrix
}
\usage{data("Results")}
\format{
  The format is:
List of 2
 $ Output   :'data.frame':	14 obs. of  15 variables:
  ..$ SampleID       : chr [1:14] "1772-062-248_A01" "1772-062-248_A02" "1772-062-248_A03" "1772-062-248_A04" ...
  ..$ X              : num [1:14] 259 261 262 261 261 258 259 189 195 194 ...
  ..$ Y              : num [1:14] 367 335 368 369 335 367 336 278 250 284 ...
  ..$ Size           : num [1:14] 31 49 19 152 141 43 59 15 49 32 ...
  ..$ Estimation.Type: chr [1:14] "Fluorescence-based" "Chip.Pattern-based" "Fluorescence-based" "Fluorescence-based" ...
  ..$ fore_Green     : num [1:14] 48.4 18.4 26.2 45.7 32.6 ...
  ..$ back_Green     : num [1:14] 17.2 16.8 16.6 16.9 17.1 ...
  ..$ fore_Red       : num [1:14] 219.1 19.8 86.5 18.4 48 ...
  ..$ back_Red       : num [1:14] 17.5 17.8 17.5 18.1 18 ...
  ..$ Signal-to-Noise: Factor w/ 14 levels "0.141617112045591",..: 12 1 8 5 4 7 10 6 3 11 ...
  ..$ Pvalue         : Factor w/ 14 levels "0.000360562783835169",..: 13 7 14 11 9 12 8 1 5 10 ...
  ..$ FDR            : Factor w/ 12 levels "0.0001684989130688",..: 9 6 1 7 11 10 12 2 6 8 ...
  ..$ Out.Index      : Factor w/ 2 levels "confidence","contamination": 1 2 1 1 1 1 1 1 1 1 ...
  ..$ Other.Spots    : Factor w/ 3 levels "0","X = 128, Y = 358 (Green) | X = 191, Y = 277 (Red)",..: 1 1 3 1 1 1 1 1 1 1 ...
  ..$ Cells          : num [1:14] 1 0 1 1 1 1 1 1 0 1 ...
 $ dateIndex: chr "WedApr611:21:282016"
}
\value{
example intermediates
}
\keyword{datasets}