% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{revDDHFinput}
\alias{revDDHFinput}
\title{revDDHFinput}
\usage{
revDDHFinput(data, hft)
}
\arguments{
\item{data}{Numeric Data matrix. A data matrix of DDHF transformed data.}
}
\value{
The reverted data
}
\description{
It reverts the DDHF sorted fluorescence signals into the original sorting.
}
\keyword{internal}

