% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{signal.from.matrix}
\alias{signal.from.matrix}
\title{signal.from.matrix}
\usage{
signal.from.matrix(coords, mat)
}
\arguments{
\item{coords}{Numeric vector. A vector of 2d coordinates for the spot center.}

\item{mat}{Data Matrix. The data matrix which generates the image of interest.}
}
\value{
Some values of interest
}
\description{
A helper to simulate spots in a image.
}
\keyword{internal}

