% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{significantSignal}
\alias{significantSignal}
\title{significantSignal}
\usage{
significantSignal(centerR, centerG, minDiff, chaImgs)
}
\arguments{
\item{centerR}{Data matrix. The location statistics of one channel.}

\item{centerG}{Data matrix. The location statistics of the other channel.}

\item{minDiff}{Float. the mu_hat of the H0: image-to-noise ratio =
log(foreground_signal) - log(background_signal) = mu_hat. Rejection of H0
implies that the identified spot is brighter than background. Default is 0.5.}

\item{chaImgs}{List. A list of the channel images (data matrices) of a sample.}
}
\value{
P-values and test statistics for both channels
}
\description{
It tests whether the foreground signal is statistically higher than the background.
}
\keyword{internal}

