% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{sortCentroids}
\alias{sortCentroids}
\title{sortCentroids}
\usage{
sortCentroids(data, type)
}
\arguments{
\item{data}{Data matrix. A matrix of centroids with their trigonometric function values.}

\item{type}{Character string. A user-defined value that characterizes the cell progression dynamics.
It can be either "clockwise" or "anticlockwise" depending on how the path is expected
to proceed.}
}
\value{
A matrix of sorted centroids
}
\description{
A helper funcion to sort the centroids .
}
\keyword{internal}

