% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{spot.simulator}
\alias{spot.simulator}
\title{spot.simulator}
\usage{
spot.simulator(location, size, average.signal, dimension)
}
\arguments{
\item{location}{Numeric vector. A vector of 2d coordinates for the spot center.}

\item{size}{Integer. The number of pixels the spot consists of.}

\item{average.signal}{Float. The parameter of the poisson distribution to generate the}

\item{dimension}{Numeric vector. The image dimensions.}
}
\value{
The image with the generated spot(s)
}
\description{
A helper to simulate spots in a image.
}
\keyword{internal}

