% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{SpotbyStrLines}
\alias{SpotbyStrLines}
\title{SpotbyStrLines}
\usage{
SpotbyStrLines(binImg, pattern.search, stats)
}
\arguments{
\item{binImg}{List. The binary segmented image data in each channel.}

\item{pattern.search}{Integer. A cutoff to find horizontal and vertical lines on the chip.}

\item{stats}{List. The estimated parameters of the BF image modelling.}
}
\value{
A series of spot location estimates
}
\description{
It estimates the spot location using the BF image modeling parameters
}
\keyword{internal}

