% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{spotCenter}
\alias{spotCenter}
\title{spotCenter}
\usage{
spotCenter(img, foregroundCut, howbig, ImgLimits)
}
\arguments{
\item{img}{Data matrix. The channel image data.}

\item{foregroundCut}{Numeric vector. The binary segmentation image analysis cutoffs
for normalized image data. Pixels with normalized signals higher than the cutoff
belong to foreground.}

\item{howbig}{Integer. An user defined value of the minimum expected spot size.}

\item{ImgLimits}{a cutoff that determines where the spot is supposed to be found.}
}
\value{
A list of spot coordinate estimates
}
\description{
It estimates a series of spot statistics on each channel.
}
\keyword{internal}

