\name{step1}
\alias{step1}
\docType{data}
\title{
step1
}
\description{
Example output of the createFluo function
}
\usage{data("step1")}
\format{
  The format is:
List of 7
 $ index     : int [1:246] 1 2 3 4 5 6 7 8 9 10 ...
 $ RGexprs   :'data.frame':	246 obs. of  4 variables:
  ..$ fore_Green: num [1:246] 48.4 26.2 45.7 34 24 ...
  ..$ back_Green: num [1:246] 17.2 17 16.8 17.1 16.8 ...
  ..$ fore_Red  : num [1:246] 219.1 86.5 18.4 84.4 104.3 ...
  ..$ back_Red  : num [1:246] 17.3 17 18 17.7 17.5 ...
 $ samples   : chr [1:246] "1772-062-248_A01" "1772-062-248_A03" "1772-062-248_A04" "1772-062-248_A06" ...
 $ batch     : chr [1:246, 1:2] "1772-062-248" "1772-062-248" "1772-062-248" "1772-062-248" ...
 $ Size      : num [1:246] 31 19 152 43 59 21 72 81 31 56 ...
 $ image.type: chr [1:3] "BF" "Green" "Red"
 $ dateIndex : chr "WedMar2313:29:522016"
}
\value{
example intermediates
}
\keyword{datasets}
