\name{step2}
\alias{step2}
\docType{data}
\title{
step2
}
\description{
Example output of the Fluo_adjustment function
}
\usage{data("step2")}
\format{
  The format is:
List of 3
 $ General             :List of 15
  ..$ index                : int [1:246] 1 2 3 4 5 6 7 8 9 10 ...
  ..$ samples              : chr [1:246] "1772-062-248_A01" "1772-062-248_A03" "1772-062-248_A04" "1772-062-248_A06" ...
  ..$ batch                : chr [1:246, 1:2] "1772-062-248" "1772-062-248" "1772-062-248" "1772-062-248" ...
  ..$ Size                 : num [1:246] 31 19 152 43 59 21 72 81 31 56 ...
  ..$ RGexprs              :'data.frame':	246 obs. of  4 variables:
  .. ..$ fore_Green: num [1:246] 48.4 26.2 45.7 34 24 ...
  .. ..$ back_Green: num [1:246] 17.2 17 16.8 17.1 16.8 ...
  .. ..$ fore_Red  : num [1:246] 219.1 86.5 18.4 84.4 104.3 ...
  .. ..$ back_Red  : num [1:246] 17.3 17 18 17.7 17.5 ...
  ..$ exprs                : num [1:246, 1:2] 35.3 13.2 32.9 20.9 11.3 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Green" "Red"
  ..$ image.type           : chr [1:3] "BF" "Green" "Red"
  ..$ dateIndex            : chr "WedMar2313:29:522016"
  ..$ single.batch.analysis: num 5
  ..$ BGmethod             : chr "normexp"
  ..$ maxMix               : num 3
  ..$ prior.pi             : num 0.1
  ..$ flex.reps            : num 5
  ..$ flexmethod           : chr "BIC"
  ..$ RNG                  : NULL
 $ Summarized_estimates:List of 3
  ..$ corrected.exprs            : num [1:246, 1:2] 37 11.8 34.4 20.5 9.7 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Green" "Red"
  ..$ corrected.transformed.exprs: num [1:246, 1:2] 3.61 2.47 3.54 3.02 2.27 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "Green" "Red"
  ..$ allResults                 :List of 2
  .. ..$ corrected.exprs            : chr [1:247, 1:10] "Ref=1" "40.7543386916705" "13.2747695514339" "37.8669946332554" ...
  .. ..$ corrected.transformed.exprs: chr [1:247, 1:10] "Ref=1" "3.70756230498048" "2.58586520738665" "3.63407987867093" ...
 $ Batch_estimates     :List of 5
  ..$ Batch1:List of 15
  .. ..$ corrected.exprs             : num [1:246, 1:2] 40.8 13.3 37.9 22.8 10.8 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ corrected.transformed.exprs : num [1:246, 1:2] 3.71 2.59 3.63 3.13 2.38 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ mixes.Green                 : num [1:246] 1 1 1 1 1 1 2 1 2 1 ...
  .. ..$ mixes.Red                   : num [1:246] 3 3 1 3 3 3 3 3 3 3 ...
  .. ..$ Batch.Green.est             : chr [1:11, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Batch)2" ...
  .. ..$ BatchRed.est                : chr [1:16, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Comp)3" ...
  .. ..$ fitted.values               : num [1:246, 1:2] 2.7 2.7 2.7 2.7 2.7 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ transformation              : chr "log"
  .. ..$ model.residuals             : num [1:246, 1:2] 0.865 -0.114 0.797 0.34 -0.277 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ model.standardized.residuals: num [1:246, 1:2] 1.746 -0.231 1.609 0.687 -0.559 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ residual.statistics         : chr [1:4, 1:7] "" "Green" "Red" "Green & Red" ...
  .. ..$ lpar                        : NULL
  .. ..$ design.Green                : num [1:246, 1:10] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:10] "(Intercept)" "factor(Comp)2" "factor(Batch)2" "factor(Batch)3" ...
  .. .. ..- attr(*, "assign")= int [1:10] 0 1 2 2 2 2 3 3 3 3
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ design.Red                  : num [1:246, 1:15] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:15] "(Intercept)" "factor(Comp)2" "factor(Comp)3" "factor(Batch)2" ...
  .. .. ..- attr(*, "assign")= int [1:15] 0 1 1 2 2 2 2 3 3 3 ...
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ reference                   : int 1
  ..$ Batch2:List of 15
  .. ..$ corrected.exprs             : num [1:246, 1:2] 34.33 9.63 31.69 17.88 8.08 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ corrected.transformed.exprs : num [1:246, 1:2] 3.54 2.26 3.46 2.88 2.09 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ mixes.Green                 : num [1:246] 1 1 1 1 1 1 1 1 2 1 ...
  .. ..$ mixes.Red                   : num [1:246] 3 3 1 3 3 3 3 3 3 3 ...
  .. ..$ Batch.Green.est             : chr [1:11, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Batch)1" ...
  .. ..$ BatchRed.est                : chr [1:16, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Comp)3" ...
  .. ..$ fitted.values               : num [1:246, 1:2] 2.85 2.85 2.85 2.85 2.85 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ transformation              : chr "log"
  .. ..$ model.residuals             : num [1:246, 1:2] 0.717 -0.262 0.649 0.192 -0.425 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ model.standardized.residuals: num [1:246, 1:2] 1.345 -0.492 1.218 0.361 -0.797 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ residual.statistics         : chr [1:4, 1:7] "" "Green" "Red" "Green & Red" ...
  .. ..$ lpar                        : NULL
  .. ..$ design.Green                : num [1:246, 1:10] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:10] "(Intercept)" "factor(Comp)2" "factor(Batch)1" "factor(Batch)3" ...
  .. .. ..- attr(*, "assign")= int [1:10] 0 1 2 2 2 2 3 3 3 3
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ design.Red                  : num [1:246, 1:15] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:15] "(Intercept)" "factor(Comp)2" "factor(Comp)3" "factor(Batch)1" ...
  .. .. ..- attr(*, "assign")= int [1:15] 0 1 1 2 2 2 2 3 3 3 ...
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ reference                   : int 2
  ..$ Batch3:List of 15
  .. ..$ corrected.exprs             : num [1:246, 1:2] 41.7 12.6 38.7 22.7 10.1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ corrected.transformed.exprs : num [1:246, 1:2] 3.73 2.53 3.66 3.12 2.31 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ mixes.Green                 : num [1:246] 1 1 1 1 1 1 2 1 2 1 ...
  .. ..$ mixes.Red                   : num [1:246] 3 3 1 3 3 3 3 3 3 3 ...
  .. ..$ Batch.Green.est             : chr [1:11, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Batch)1" ...
  .. ..$ BatchRed.est                : chr [1:16, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Comp)3" ...
  .. ..$ fitted.values               : num [1:246, 1:2] 2.7 2.7 2.7 2.7 2.7 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ transformation              : chr "log"
  .. ..$ model.residuals             : num [1:246, 1:2] 0.865 -0.114 0.797 0.34 -0.277 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ model.standardized.residuals: num [1:246, 1:2] 1.749 -0.231 1.612 0.688 -0.56 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ residual.statistics         : chr [1:4, 1:7] "" "Green" "Red" "Green & Red" ...
  .. ..$ lpar                        : NULL
  .. ..$ design.Green                : num [1:246, 1:10] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:10] "(Intercept)" "factor(Comp)2" "factor(Batch)1" "factor(Batch)2" ...
  .. .. ..- attr(*, "assign")= int [1:10] 0 1 2 2 2 2 3 3 3 3
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ design.Red                  : num [1:246, 1:15] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:15] "(Intercept)" "factor(Comp)2" "factor(Comp)3" "factor(Batch)1" ...
  .. .. ..- attr(*, "assign")= int [1:15] 0 1 1 2 2 2 2 3 3 3 ...
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ reference                   : int 3
  ..$ Batch4:List of 15
  .. ..$ corrected.exprs             : num [1:246, 1:2] 34.67 11.85 32.27 19.75 9.81 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ corrected.transformed.exprs : num [1:246, 1:2] 3.55 2.47 3.47 2.98 2.28 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ mixes.Green                 : num [1:246] 1 1 1 1 1 1 2 1 2 1 ...
  .. ..$ mixes.Red                   : num [1:246] 3 3 1 3 3 3 3 3 3 3 ...
  .. ..$ Batch.Green.est             : chr [1:11, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Batch)1" ...
  .. ..$ BatchRed.est                : chr [1:16, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Comp)3" ...
  .. ..$ fitted.values               : num [1:246, 1:2] 2.7 2.7 2.7 2.7 2.7 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ transformation              : chr "log"
  .. ..$ model.residuals             : num [1:246, 1:2] 0.865 -0.114 0.797 0.34 -0.277 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ model.standardized.residuals: num [1:246, 1:2] 1.741 -0.23 1.605 0.685 -0.557 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ residual.statistics         : chr [1:4, 1:7] "" "Green" "Red" "Green & Red" ...
  .. ..$ lpar                        : NULL
  .. ..$ design.Green                : num [1:246, 1:10] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:10] "(Intercept)" "factor(Comp)2" "factor(Batch)1" "factor(Batch)2" ...
  .. .. ..- attr(*, "assign")= int [1:10] 0 1 2 2 2 2 3 3 3 3
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ design.Red                  : num [1:246, 1:15] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:15] "(Intercept)" "factor(Comp)2" "factor(Comp)3" "factor(Batch)1" ...
  .. .. ..- attr(*, "assign")= int [1:15] 0 1 1 2 2 2 2 3 3 3 ...
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ reference                   : int 4
  ..$ Batch5:List of 17
  .. ..$ corrected.exprs             : num [1:246, 1:2] 34.3 11.96 31.95 19.7 9.95 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ corrected.transformed.exprs : num [1:246, 1:2] 3.54 2.48 3.46 2.98 2.3 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ mixes.Green                 : num [1:246] 1 1 1 1 1 1 2 1 2 1 ...
  .. ..$ mixes.Red                   : num [1:246] 2 2 1 2 2 2 2 2 2 2 ...
  .. ..$ Batch.Green.est             : chr [1:11, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Batch)1" ...
  .. ..$ BatchRed.est                : chr [1:11, 1:4] "" "(Intercept)" "factor(Comp)2" "factor(Batch)1" ...
  .. ..$ fitted.values               : num [1:246, 1:2] 2.7 2.7 2.7 2.7 2.7 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ transformation              : chr "log"
  .. ..$ model.residuals             : num [1:246, 1:2] 0.865 -0.114 0.797 0.34 -0.277 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ model.standardized.residuals: num [1:246, 1:2] 1.749 -0.231 1.612 0.688 -0.56 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:2] "Green" "Red"
  .. ..$ residual.statistics         : chr [1:4, 1:7] "" "Green" "Red" "Green & Red" ...
  .. ..$ lpar                        : NULL
  .. ..$ design.Green                : num [1:246, 1:10] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:10] "(Intercept)" "factor(Comp)2" "factor(Batch)1" "factor(Batch)2" ...
  .. .. ..- attr(*, "assign")= int [1:10] 0 1 2 2 2 2 3 3 3 3
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ design.Red                  : num [1:246, 1:10] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:246] "1" "2" "3" "4" ...
  .. .. .. ..$ : chr [1:10] "(Intercept)" "factor(Comp)2" "factor(Batch)1" "factor(Batch)2" ...
  .. .. ..- attr(*, "assign")= int [1:10] 0 1 2 2 2 2 3 3 3 3
  .. .. ..- attr(*, "contrasts")=List of 2
  .. .. .. ..$ factor(Comp) : chr "contr.treatment"
  .. .. .. ..$ factor(Batch): chr "contr.treatment"
  .. ..$ reference                   : int 5
  .. ..$ Green.contrasts             : chr [1:21, 1:7] "Channel" "CH1" "CH1" "CH1" ...
  .. ..$ Red.contrasts               : chr [1:21, 1:7] "Channel" "CH2" "CH2" "CH2" ...
}
\value{
example intermediates
}
\keyword{datasets}
