% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{straightColLines}
\alias{straightColLines}
\title{straightColLines}
\usage{
straightColLines(img, pattern.search, cut = seq(0.3, 1.5, 0.02), ImgLimits,
  chip.type)
}
\arguments{
\item{img}{Data matrix. The BF image data matrix.}

\item{pattern.search}{Integer. A cutoff to find horizontal and vertical lines on the chip.}

\item{cut}{Integer. A normalized signal cutoff above which a pixel is considered as foreground.}

\item{ImgLimits}{Integer. It instructs the algorithm to find spots in a certain central image area.
For example, for a 512 x 512 image with cutSides = 50, it will search for spots in the central
area [ImgLimits:(512-ImgLimits),ImgLimits:(512-ImgLimits)] of the image matrix.}

\item{chip.type}{Character string. It specifies the type of Fluidigm chip to be analyzed.}
}
\value{
Estimated vertical BF characteristic lines.
}
\description{
It identifies the vertical BF image characteristic lines.
}
\keyword{internal}

