% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{straightRowLines}
\alias{straightRowLines}
\title{straightRowLines}
\usage{
straightRowLines(colData, pattern.search, ImgLimits)
}
\arguments{
\item{colData}{List. The data to be analyzed.}

\item{pattern.search}{Integer. A cutoff to find horizontal and vertical lines on the chip}

\item{ImgLimits}{Integer. It instructs the algorithm to find spots in a certain central image area.
For example, for a 512 x 512 image with cutSides = 50, it will search for spots in the central
area [ImgLimits:(512-ImgLimits),ImgLimits:(512-ImgLimits)] of the image matrix.}
}
\value{
The estimated horizontal BF characteristic lines
}
\description{
It identifies the horizontal BF image characteristic lines.
}
\keyword{internal}

