% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{subsetAnalysis}
\alias{subsetAnalysis}
\title{subsetAnalysis}
\usage{
subsetAnalysis(files, sub)
}
\arguments{
\item{files}{List. A list of all BF and channel file names.}

\item{sub}{Character string. A vector of the file names to be read.}
}
\value{
A list of files to be analyzed
}
\description{
It reads a subset of the original file names.
}
\keyword{internal}

