% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{transformFluo}
\alias{transformFluo}
\title{transformFluo}
\usage{
transformFluo(data, method)
}
\arguments{
\item{data}{List. A list of adjusted fluorescence signals in both channels. Typically,
the output of orderFluo().}

\item{method}{Character string. The variance stabilizing method. One of "DDHFmv" or "log".}
}
\value{
A list with the adjusted fluorescence signals and the transformed adjusted fluorescence signals
}
\description{
It performs the variance stabilizing transformation of the flurescence signals.
}
\keyword{internal}

