% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{trigofun}
\alias{trigofun}
\title{trigofun}
\usage{
trigofun(data)
}
\arguments{
\item{data}{Data matrix. A matrix of fluorescence centroids.}
}
\value{
The centroids and their trigonometric function values
}
\description{
It computes the sin,cos and tan trigonometric functions for a number of points (2-dimensional fluorescence centoids)
  relative to the start of the axes.
}
\keyword{internal}

