% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{unnormalizeC01}
\alias{unnormalizeC01}
\title{unnormalizeC01}
\usage{
unnormalizeC01(data, bits)
}
\arguments{
\item{data}{Data matrix. The matrix of the normalized image signals.}

\item{bits}{Numeric. The image bits.}
}
\value{
A matrix of unnormalized image signals
}
\description{
It calculates the unnormalized signal of a normalized image. The normalization has been originally performed
  as x[i,j]/bits where x[i,j] is the i,j element of the signal matrix x and bits is the value of the bits.
}
\keyword{internal}

