% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{updateCentroids}
\alias{updateCentroids}
\title{updateCentroids}
\usage{
updateCentroids(data, centroidTable)
}
\arguments{
\item{data}{Data matrix. A matrix of appropriately transformed fluorescence signals .}

\item{centroidTable}{Data matrix. A previously estimated centroids table to be updated.}
}
\value{
The updated centroids table
}
\description{
It updates the centroids of the clusters that are re-estimated by change-point analysis.
}
\keyword{internal}

