% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_NBE.R
\name{updateCentroidsPaths}
\alias{updateCentroidsPaths}
\title{updateCentroidsPaths}
\usage{
updateCentroidsPaths(data, estimates, path.type)
}
\arguments{
\item{data}{List. A list of adjusted fluorescence signals.}

\item{estimates}{List. A list of sorted Ch2-Ch3 transformed fluorescence signals with
their assocuated change-points.}

\item{path.type}{Character vector. A user-defined vector that characterizes the cell progression dynamics.
The first element can be either "circular" or "A2Z" or "other". If "circular" the path progression is
assummed to exhibit a circle-like behavior. If "A2Z" the path is assumed to have a well-defined start
and a well-defined end point (e.g. a linear progression). If "other" the progression is assumed to be
arbitrary without an obvious directionality.}
}
\value{
A list of adjusted fluorescence signals and the updated path after the change-point analysis
}
\description{
It updates the path sorted clusters after re-estimation by change-point analysis.
}
\keyword{internal}

