% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_DDHF.R
\name{which.min.diff}
\alias{which.min.diff}
\title{which.min.diff}
\usage{
which.min.diff(a, vect)
}
\arguments{
\item{a, vector}{Appropriate vectors for analysis}
}
\value{
Preliminary DDHF results
}
\description{
A helper for DDHF
}
\keyword{internal}

