% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fluo_est.R
\name{zoomInBF}
\alias{zoomInBF}
\title{zoomInBF}
\usage{
zoomInBF(img, pattern.search, ImgLimits, chip.type)
}
\arguments{
\item{img}{Data matrix. The BF image data.}

\item{pattern.search}{Integer. A cutoff to find horizontal and vertical lines on the chip.}

\item{ImgLimits}{Integer. It instructs the algorithm to find spots in a certain central image area.
For example, for a 512 x 512 image with cutSides = 50, it will search for spots in the central
area [ImgLimits:(512-ImgLimits),ImgLimits:(512-ImgLimits)] of the image matrix.}

\item{chip.type}{Character string. It specifies the type of Fluidigm chip to be analyzed.}
}
\value{
The locations of the straight lines on the chip
}
\description{
It estimates the chip characterisrics for BF image modelling
}
\keyword{internal}

