\name{Results}
\alias{Results}
\docType{data}
\title{
Results
}
\description{
Example output from LocationMatrix call
}
\usage{data("Results")}
\format{
  The format is:
List of 2
 $ Output   :'data.frame':	14 obs. of  15 variables:
  ..$ SampleID       : chr [1:14] "1772-062-248_A01" "1772-062-248_A02" "1772-062-248_A03" "1772-062-248_A04" ...
  ..$ X              : num [1:14] 259 491 262 261 261 258 259 189 498 194 ...
  ..$ Y              : num [1:14] 367 219 368 369 335 367 336 278 20 284 ...
  ..$ Size           : num [1:14] 31 49 19 152 141 43 59 15 49 32 ...
  ..$ Estimation.Type: chr [1:14] "Fluorescence-based" "Fluorescence-based" "Fluorescence-based" "Fluorescence-based" ...
  ..$ fore_Green     : num [1:14] 48.4 36 26.2 45.7 32.6 ...
  ..$ back_Green     : num [1:14] 17.1 17.3 16.7 16.8 17.1 ...
  ..$ fore_Red       : num [1:14] 219.1 27.6 86.5 18.4 48 ...
  ..$ back_Red       : num [1:14] 17.4 18.8 17.5 18 18.1 ...
  ..$ Signal-to-Noise: Factor w/ 14 levels "0.150964535550184",..: 12 3 8 5 4 7 9 6 2 11 ...
  ..$ Pvalue         : Factor w/ 14 levels "0.000360562783835169",..: 13 3 14 11 9 12 8 1 6 10 ...
  ..$ FDR            : Factor w/ 13 levels "0.000181460367920246",..: 10 4 1 8 12 11 13 2 7 9 ...
  ..$ Out.Index      : Factor w/ 1 level "confidence": 1 1 1 1 1 1 1 1 1 1 ...
  ..$ Other.Spots    : Factor w/ 3 levels "0","X = 128, Y = 358 (Green) | X = 191, Y = 277 (Red)",..: 1 1 3 1 1 1 1 1 1 1 ...
  ..$ Cells          : num [1:14] 1 1 1 1 1 1 1 1 0 1 ...
 $ dateIndex: chr "WedMar2313:16:562016"
}
\value{
example intermediates
}
\keyword{datasets}
